MDEV-16889: Spider Crash mysqld got exception 0xc0000005
The SELECT with the INNER JOIN is executed with one of the two tables being
optimized as a constant table, which is pre-read.  Spider nevertheless attempts
to push down the join to the data node.  The crash occurs because the constant
table is excluded from the optimized query that Spider attempts to push down.

In order for Spider to be able to push down a join, all of the tables involved
in the join need to be included in the optimized query that Spider pushes down.
When any of the tables involved in the join is a constant table, it is excluded
from the optimized query that Spider attempts to push down.  I have therefore
fixed the problem by preventing Spider from pushing down queries that include
tables that have been optimized as constant tables.

Author:
  Jacob Mathew.

Reviewer:
  Kentoku Shiba.